# polys: Replace algorithm of primitive_element for ex=True