--- a/sympy/polys/numberfields.py
+++ b/sympy/polys/numberfields.py
@@ -842,12 +842,7 @@ def primitive_element(extension, x=None, *, ex=False, coeffs=_coeffs_generator, 
 
     if not ex:
         gen, coeffs = extension[0], [1]
-        # XXX when minimal_polynomial is extended to work
-        # with AlgebraicNumbers this test can be removed
-        if isinstance(gen, AlgebraicNumber):
-            g = gen.minpoly.replace(x)
-        else:
-            g = minimal_polynomial(gen, x, polys=True)
+        g = minimal_polynomial(gen, x, polys=True)
         for ext in extension[1:]:
             _, factors = factor_list(g, extension=ext)
             g = _choose_factor(factors, x, gen)