--- a/sympy/polys/tests/test_numberfields.py
+++ b/sympy/polys/tests/test_numberfields.py
@@ -18,6 +18,7 @@ from sympy.polys.numberfields import (
     isolate, IntervalPrinter,
 )
 
+from sympy.polys.partfrac import apart
 from sympy.polys.polyerrors import (
     IsomorphismFailed,
     NotAlgebraic,
@@ -796,3 +797,12 @@ def test_issue_19760():
     for comp in (True, False):
         mp = Poly(minimal_polynomial(e, compose=comp))
         assert mp(x) == mp_expected, "minimal_polynomial(e, compose=%s) = %s; %s expected" % (comp, mp(x), mp_expected)
+
+
+def test_issue_20163():
+    assert apart(1/(x**6+1), extension=[sqrt(3), I]) == \
+        (sqrt(3) + I)/(2*x + sqrt(3) + I)/6 + \
+        (sqrt(3) - I)/(2*x + sqrt(3) - I)/6 - \
+        (sqrt(3) - I)/(2*x - sqrt(3) + I)/6 - \
+        (sqrt(3) + I)/(2*x - sqrt(3) - I)/6 + \
+        I/(x + I)/6 - I/(x - I)/6