<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->
Fixes #20163

#### Brief description of what is fixed or changed


The algorithm currently used by primitive_element for ex=True is
based on the `groebner` function. That does not work reliably if
the input extension elements are not linearly disjoint (see e.g.
#14117). This commit implements the same iterative method
that is used for the ex=False case extended
by including representations of field elements.


#### Other comments


#### Release Notes

<!-- Write the release notes for this release below. See
https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more information
on how to write release notes. The bot will check your release notes
automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
- polys
  - numberfields: implemented new algorithm for `primitive_element` in case `ex=True`.
<!-- END RELEASE NOTES -->