@oscarbenjamin 2020-10-17
Let me try to understand how this works...

We want to construct an algebraic number field containing some expressions like:
```julia
In [4]: exts = [sqrt(3)+I, -sqrt(3)+I, sqrt(3)-I, -sqrt(3)-I, -2*I, 2*I]                                                                       

In [5]: construct_domain(exts, extension=True)                                                                                                 
^C---------------------------------------------------------------------------
KeyboardInterrupt
```
Clearly these extensions are all generated by `sqrt(3)` and `I` over `QQ`. In fact by the primitive element theorem we can have a single generator like `sqrt(3) + I` so it can be `QQ<sqrt(3) + I>` with minimal polynomial:
```julia
In [6]: minimal_polynomial(sqrt(3) + I)                                                                                                        
Out[6]: 
 4      2     
x  - 4⋅x  + 16
```
On master this primitive element is calculated by computing the minimal polynomials for each of `exts` and then computing their Groebner basis which can be very slow.

In this PR it is instead computed without the Groebner basis and succeeds in finding a primitive element `-sqrt(3)-I` (the primitive element found is non-deterministic):
```julia
In [1]: exts = [sqrt(3)+I, -sqrt(3)+I, sqrt(3)-I, -sqrt(3)-I, -2*I, 2*I]                                                                       

In [2]: construct_domain(exts, extension=True)                                                                                                 
Out[2]: 
(QQ<-sqrt(3) - I>,
 [ANP([mpq(-1,1), mpq(0,1)], [mpq(1,1), mpq(0,1), mpq(-4,1), mpq(0,1), mpq(16,1)], QQ),
  ANP([mpq(-1,4), mpq(0,1), mpq(1,1), mpq(0,1)], [mpq(1,1), mpq(0,1), mpq(-4,1), mpq(0,1), mpq(16,1)], QQ),
  ANP([mpq(1,4), mpq(0,1), mpq(-1,1), mpq(0,1)], [mpq(1,1), mpq(0,1), mpq(-4,1), mpq(0,1), mpq(16,1)], QQ),
  ANP([mpq(1,1), mpq(0,1)], [mpq(1,1), mpq(0,1), mpq(-4,1), mpq(0,1), mpq(16,1)], QQ),
  ANP([mpq(1,4), mpq(0,1), mpq(0,1), mpq(0,1)], [mpq(1,1), mpq(0,1), mpq(-4,1), mpq(0,1), mpq(16,1)], QQ),
  ANP([mpq(-1,4), mpq(0,1), mpq(0,1), mpq(0,1)], [mpq(1,1), mpq(0,1), mpq(-4,1), mpq(0,1), mpq(16,1)], QQ)])
```

The method used here receives a slightly filtered list of extensions:
```python
[I, sqrt(3) + I, sqrt(3) - I, -sqrt(3) - I, -sqrt(3) + I]
```
This has combined `2*I` and `-2*I` and removed the rational factor giving just `I`. I'm not sure where that preprocessing happened but would it not have been straight-forward to process the Adds as well to end up with just `[I, sqrt(3)]`?

In any case the method proceeds by computing the minimal polynomial for each element over `QQ` and then attempting to combine those into a single extension field. In the first iteration we take `QQ<exts[0]>` and `QQ<exts[1]>` and then try to factor one minimal polynomial over the other extension field. I guess if that succeeds then the minimal polynomial for one is reducible over the other field and so it's generators are not needed:
```julia
In [7]: ext1 = I                                                                                                                               

In [8]: ext2 = sqrt(3) + I                                                                                                                     

In [9]: f = minimal_polynomial(ext1, x, polys=True)                                                                                            

In [10]: K = QQ.algebraic_field((f, ext1))                                                                                                     

In [11]: g = minimal_polynomial(ext2, x, polys=True)                                                                                           

In [12]: L = QQ.algebraic_field((f, ext2))                                                                                                     

In [13]: f                                                                                                                                     
Out[13]: Poly(x**2 + 1, x, domain='QQ')

In [14]: g                                                                                                                                     
Out[14]: Poly(x**4 - 4*x**2 + 16, x, domain='QQ')

In [15]: K                                                                                                                                     
Out[15]: QQ<I>

In [16]: L                                                                                                                                     
Out[16]: QQ<sqrt(3) + I>

In [17]: factor_list(f, domain=L)                                                                                                              
Out[17]: (1, [(Poly(x - sqrt(3) - I, x, domain='QQ<sqrt(3) + I>'), 1), (Poly(x + sqrt(3) + I, x, domain='QQ<sqrt(3) + I>'), 1)])
```
I'm not quite sure what the `sqf_norm` and the `gcd` are doing but I guess they are trying to normalise the result in some way.

It seems that the loop will attempt to convert all reps to a new field on each iteration. Perhaps it should be checked if the new K is actually different from the previous K.

Why is the method non-deterministic? Maybe the ordering of `extension` comes from a `set` somewhere.

Is it possible to preprocess this and extract the unique non-rational generators from the Adds?

Presumably this does not fix #14177 so what are examples that still don't work? The OP example seems to work with this change.