@jksuom 2020-10-17
This works by adjoining one `ext` element at a time to Q. The primitive element of the growing extension is denoted by `gen`. At each step an `ext` is adjoined to `K = Q(gen)`. `K(ext)` can then be generated by a rational linear combination `gen + s*ext` for all but a finite number of coefficients `s`. It is the job of `sqf_norm` to find a suitable `s`. It tries `s = 0, 1, 2, ...` and stops at the first success. (So, essentially, this is quite deterministic.)

For a polynomial `f` with domain `Q(ext)`, `f.sqf_norm()` returns `s, g, r` where `g(x) = f(x - s*ext)` (there is a typo in the docstring), `r` is its norm from `Q(ext)` to `Q` and `s` has been chosen such that `r` is squarefree. In the algorithm, `f` will be the minimal polynomial of `gen` over `Q(ext)`, hence irreducible. Then `g` will also be irreducible and its norm will be a power of an irreducible polynomial, so also irreducible being squarefree. Hence it will be a minimal polynomial of the new generator `gen + s*ext`.

> Perhaps it should be checked if the new K is actually different from the previous K.

This would be the case `s == 0`. It could be tested but I'm not sure that it would be worthwhile.
It will only happen with redundant `ext` elements.

The role of `gcd` may look obscure. The polynomial `g` is an algebraic relation of the new `gen` over `Q(ext)`. It is first transformed to a relation `h` of `ext` over `Q(gen)`. We know that there is a linear relation as `ext` is an element of `Q(gen)`. Taking the `gcd` of `h` with the minimal polynomial of `ext` over `Q` will create such a relation.