@oscarbenjamin 2020-10-17
There is definitely something non-deterministic. Given `t.py`:
```python
from sympy import *

exts = [sqrt(3)+I, -sqrt(3)+I, sqrt(3)-I, -sqrt(3)-I, -2*I, 2*I]
print(construct_domain(exts, extension=True)[0])
```
I get different results on different runs:
```console
$ for n in `seq 1 10`; do python t.py ; done
QQ<sqrt(3) + I>
QQ<-2*sqrt(3) - I>
QQ<sqrt(3) - I>
QQ<-2*sqrt(3) - I>
QQ<-sqrt(3) - I>
QQ<-sqrt(3) - I>
QQ<-sqrt(3) + 2*I>
QQ<-sqrt(3) - I>
QQ<-sqrt(3) - I>
QQ<sqrt(3) + I>
```