@oscarbenjamin 2020-12-15
We can also improve performance with something like this to extract the elementary algebraic parts of the expression:
```diff
diff --git a/sympy/polys/constructor.py b/sympy/polys/constructor.py
index d6c5faa78e..22a4b3bf70 100644
--- a/sympy/polys/constructor.py
+++ b/sympy/polys/constructor.py
@@ -82,27 +82,24 @@ def _construct_algebraic(coeffs, opt):
     """We know that coefficients are algebraic so construct the extension. """
     from sympy.polys.numberfields import primitive_element
 
-    result, exts = [], set()
-
-    for coeff in coeffs:
-        if coeff.is_Rational:
-            coeff = (None, 0, QQ.from_sympy(coeff))
-        else:
-            a = coeff.as_coeff_add()[0]
-            coeff -= a
-
-            b = coeff.as_coeff_mul()[0]
-            coeff /= b
-
-            exts.add(coeff)
-
-            a = QQ.from_sympy(a)
-            b = QQ.from_sympy(b)
-
-            coeff = (coeff, b, a)
-
-        result.append(coeff)
+    exts = set()
+
+    def build_trees(args):
+        trees = []
+        for a in args:
+            if a.is_Rational:
+                tree = ('Q', QQ.from_sympy(a))
+            elif a.is_Add:
+                tree = ('+', build_trees(a.args))
+            elif a.is_Mul:
+                tree = ('*', build_trees(a.args))
+            else:
+                tree = ('e', a)
+                exts.add(a)
+            trees.append(tree)
+        return trees
 
+    trees = build_trees(coeffs)
     exts = list(ordered(exts))
 
     g, span, H = primitive_element(exts, ex=True, polys=True)
@@ -110,13 +107,27 @@ def _construct_algebraic(coeffs, opt):
 
     domain, g = QQ.algebraic_field((g, root)), g.rep.rep
 
-    for i, (coeff, a, b) in enumerate(result):
-        if coeff is not None:
-            coeff = a*domain.dtype.from_list(H[exts.index(coeff)], g, QQ) + b
+    exts_dom = [domain.dtype.from_list(h, g, QQ) for h in H]
+    exts_map = dict(zip(exts, exts_dom))
+
+    def convert_tree(tree):
+        op, args = tree
+        if op == 'Q':
+            return domain.dtype.from_list([args], g, QQ)
+        elif op == '+':
+            return sum((convert_tree(a) for a in args), domain.zero)
+        elif op == '*':
+            # return prod(convert(a) for a in args)
+            t = convert_tree(args[0])
+            for a in args[1:]:
+                t *= convert_tree(a)
+            return t
+        elif op == 'e':
+            return exts_map[args]
         else:
-            coeff = domain.dtype.from_list([b], g, QQ)
+            raise RuntimeError
 
-        result[i] = coeff
+    result = [convert_tree(tree) for tree in trees]
 
     return domain, result
 
diff --git a/sympy/polys/rings.py b/sympy/polys/rings.py
index 3585e8f0bf..9d7fe18871 100644
--- a/sympy/polys/rings.py
+++ b/sympy/polys/rings.py
@@ -163,10 +163,12 @@ def sring(exprs, *symbols, **options):
     reps, opt = _parallel_dict_from_expr(exprs, opt)
 
     if opt.domain is None:
-        # NOTE: this is inefficient because construct_domain() automatically
-        # performs conversion to the target domain. It shouldn't do this.
         coeffs = sum([ list(rep.values()) for rep in reps ], [])
-        opt.domain, _ = construct_domain(coeffs, opt=opt)
+
+        opt.domain, coeffs_dom = construct_domain(coeffs, opt=opt)
+
+        coeff_map = dict(zip(coeffs, coeffs_dom))
+        reps = [{m: coeff_map[c] for m, c in rep.items()} for rep in reps]
 
     _ring = PolyRing(opt.gens, opt.domain, opt.order)
     polys = list(map(_ring.from_dict, reps))
```

I tested performance with:
```python
from sympy import *

exts = [sqrt(prime(n)) for n in range(1, 10)]

nexts = 2

for n in range(1, 10):
    M = randMatrix(n, n + 1)
    for ext in exts[:nexts]:
        M += ext * randMatrix(n, n + 1)
    syms = symbols('x:%d' % n)
    eqs = list(M * Matrix(syms + (1,)))

    import time
    start = time.time()
    linsolve(eqs, syms)
    end = time.time()
    print('n =', n, ':', '%2g secs' % (end - start))
```
On master that gives:
```console
$ python time.py 
n = 1 : 3.18847 secs
^C
```

With this PR that gives:
```console
$ python time.py 
n = 1 : 0.229376 secs
n = 2 : 0.747598 secs
n = 3 : 1.62068 secs
n = 4 : 2.94355 secs
n = 5 : 3.71783 secs
n = 6 : 6.04225 secs
^C
```

With this PR and the diff above it gives:
```console
$ python time.py 
n = 1 : 0.0299489 secs
n = 2 : 0.0353289 secs
n = 3 : 0.0492551 secs
n = 4 : 0.0669062 secs
n = 5 : 0.0878379 secs
n = 6 : 0.127864 secs
n = 7 : 0.154448 secs
n = 8 : 0.201635 secs
n = 9 : 0.256783 secs
```

@jksuom do you mind if I push that on here?